# src/routes/swagger.json
{
  "openapi": "3.0.0",
  "components": {
...
    "schemas": {
...
      "Record_string.unknown_": {
        "properties": {},
        "additionalProperties": {},
        "type": "object",
        "description": "Construct a type with a set of properties K of type T"
      },
...
      "RecordId": {
        "type": "string",
        "example": "821f9b26-ad04-4f56-89b6-e2ef9c72b36e"
      },
...
      "PlaintextMessage": {
        "properties": {
          "@type": {
            "type": "string"
          },
          "@id": {
            "type": "string"
          },
          "~thread": {
            "properties": {
              "pthid": {
                "type": "string"
              },
              "thid": {
                "type": "string"
              }
            },
            "type": "object"
          },
          "messageType": {
            "type": "string"
          }
        },
        "required": [
          "@type",
          "@id",
          "messageType"
        ],
        "type": "object",
        "additionalProperties": {}
      },
...
      "HandshakeProtocol": {
        "description": "Enum values should be sorted based on order of preference. Values will be\nincluded in this order when creating out of band invitations.",
        "enum": [
          "https://didcomm.org/didexchange/1.x",
          "https://didcomm.org/connections/1.x"
        ],
        "type": "string"
      },
      "AgentMessage": {
        "$ref": "#/components/schemas/PlaintextMessage"
      },
...
      "Routing": {
        "properties": {
          "endpoints": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "recipientKey": {
            "$ref": "#/components/schemas/Key"
          },
          "routingKeys": {
            "items": {
              "$ref": "#/components/schemas/Key"
            },
            "type": "array"
          },
          "mediatorId": {
            "type": "string"
          }
        },
        "required": [
          "endpoints",
          "recipientKey",
          "routingKeys"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
      "CreateInvitationOptions": {
        "properties": {
          "label": {
            "type": "string"
          },
          "alias": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "goalCode": {
            "type": "string"
          },
          "goal": {
            "type": "string"
          },
          "handshake": {
            "type": "boolean"
          },
          "handshakeProtocols": {
            "items": {
              "$ref": "#/components/schemas/HandshakeProtocol"
            },
            "type": "array"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/AgentMessage"
            },
            "type": "array"
          },
          "multiUseInvitation": {
            "type": "boolean"
          },
          "autoAcceptConnection": {
            "type": "boolean"
          },
          "routing": {
            "$ref": "#/components/schemas/Routing"
          },
          "appendedAttachments": {
            "items": {
              "$ref": "#/components/schemas/Attachment"
            },
            "type": "array"
          },
          "invitationDid": {
            "type": "string"
          }
        },
        "type": "object",
        "additionalProperties": false
      },
      "RecipientKeyOption": {
        "properties": {
          "recipientKey": {
            "type": "string"
          }
        },
        "type": "object",
        "additionalProperties": false
      },
      "Pick_CreateLegacyInvitationConfig.Exclude_keyofCreateLegacyInvitationConfig.routing__": {
        "properties": {
          "label": {
            "type": "string"
          },
          "alias": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "multiUseInvitation": {
            "type": "boolean"
          },
          "autoAcceptConnection": {
            "type": "boolean"
          }
        },
        "type": "object",
        "description": "From T, pick a set of properties whose keys are in the union K"
      },
      "Omit_CreateLegacyInvitationConfig.routing_": {
        "$ref": "#/components/schemas/Pick_CreateLegacyInvitationConfig.Exclude_keyofCreateLegacyInvitationConfig.routing__",
        "description": "Construct a type with the properties of T except for those in type K."
      },
...
    },
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      },
      "jwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "info": {
    "title": "credo-controller",
    "version": "2.0.0",
    "description": "Rest endpoint wrapper for using your agent over HTTP",
    "license": {
      "name": "Apache-2.0"
    },
    "contact": {}
  },
  "paths": {
...
    "/didcomm/oob": {
      "get": {
        "operationId": "GetAllOutOfBandRecords",
        "responses": {
          "200": {
            "description": "OutOfBandRecord[]",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Record_string.unknown_"
                  },
                  "type": "array"
                },
                "examples": {
                  "Example 1": {
                    "value": [
                      {
                        "_tags": {
                          "invitationId": "1cbd22e4-1906-41e9-8807-83d84437f978",
                          "state": "await-response",
                          "role": "sender",
                          "recipientKeyFingerprints": [
                            "z6MktUCPZjfRJXD4GMcYuXiqX2qZ8vBw6UAYpDFiHEUfwuLj"
                          ]
                        },
                        "outOfBandInvitation": {
                          "@type": "did:sov:BzCbsNYhMrjHiqZDTUASHg;spec/out-of-band/1.1/invitation",
                          "@id": "d6472943-e5d0-4d95-8b48-790ed5a41931",
                          "label": "Aries Test Agent",
                          "accept": [
                            "didcomm/aip1",
                            "didcomm/aip2;env=rfc19"
                          ],
                          "handshake_protocols": [
                            "https://didcomm.org/didexchange/1.0",
                            "https://didcomm.org/connections/1.0"
                          ],
                          "services": [
                            {
                              "id": "#inline-0",
                              "serviceEndpoint": "https://6b77-89-20-162-146.ngrok.io",
                              "type": "did-communication",
                              "recipientKeys": [
                                "did:key:z6MkmTBHTWrvLPN8pBmUj7Ye5ww9GiacXCYMNVvpScSpf1DM"
                              ],
                              "routingKeys": []
                            }
                          ]
                        },
                        "metadata": {},
                        "id": "42a95528-0e30-4f86-a462-0efb02178b53",
                        "createdAt": "2022-01-01T00:00:00.000Z",
                        "reusable": false
                      }
                    ]
                  }
                }
              }
            },
            "headers": {
              "x-correlation-id": {
                "description": "Correlation ID for tracing across systems",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "description": "Retrieve all out of band records",
        "tags": [
          "DIDComm - Out Of Band"
        ],
        "security": [
          {
            "jwt": [
              "tenant",
              "dedicated"
            ]
          }
        ],
        "parameters": [
          {
            "description": "invitation identifier",
            "in": "query",
            "name": "invitationId",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/RecordId"
            }
          }
        ]
      }
    },
    "/didcomm/oob/{outOfBandId}": {
      "get": {
        "operationId": "GetOutOfBandRecordById",
        "responses": {
          "200": {
            "description": "OutOfBandRecord",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Record_string.unknown_"
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "_tags": {
                        "invitationId": "1cbd22e4-1906-41e9-8807-83d84437f978",
                        "state": "await-response",
                        "role": "sender",
                        "recipientKeyFingerprints": [
                          "z6MktUCPZjfRJXD4GMcYuXiqX2qZ8vBw6UAYpDFiHEUfwuLj"
                        ]
                      },
                      "outOfBandInvitation": {
                        "@type": "did:sov:BzCbsNYhMrjHiqZDTUASHg;spec/out-of-band/1.1/invitation",
                        "@id": "d6472943-e5d0-4d95-8b48-790ed5a41931",
                        "label": "Aries Test Agent",
                        "accept": [
                          "didcomm/aip1",
                          "didcomm/aip2;env=rfc19"
                        ],
                        "handshake_protocols": [
                          "https://didcomm.org/didexchange/1.0",
                          "https://didcomm.org/connections/1.0"
                        ],
                        "services": [
                          {
                            "id": "#inline-0",
                            "serviceEndpoint": "https://6b77-89-20-162-146.ngrok.io",
                            "type": "did-communication",
                            "recipientKeys": [
                              "did:key:z6MkmTBHTWrvLPN8pBmUj7Ye5ww9GiacXCYMNVvpScSpf1DM"
                            ],
                            "routingKeys": []
                          }
                        ]
                      },
                      "metadata": {},
                      "id": "42a95528-0e30-4f86-a462-0efb02178b53",
                      "createdAt": "2022-01-01T00:00:00.000Z",
                      "reusable": false
                    }
                  }
                }
              }
            },
            "headers": {
              "x-correlation-id": {
                "description": "Correlation ID for tracing across systems",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "description": "Retrieve an out of band record by id",
        "tags": [
          "DIDComm - Out Of Band"
        ],
        "security": [
          {
            "jwt": [
              "tenant",
              "dedicated"
            ]
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "outOfBandId",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RecordId"
            }
          }
        ]
      },
      "delete": {
        "operationId": "DeleteOutOfBandRecord",
        "responses": {
          "204": {
            "description": "No content",
            "headers": {
              "x-correlation-id": {
                "description": "Correlation ID for tracing across systems",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "description": "Deletes an out of band record from the repository.",
        "tags": [
          "DIDComm - Out Of Band"
        ],
        "security": [
          {
            "jwt": [
              "tenant",
              "dedicated"
            ]
          }
        ],
        "parameters": [
          {
            "description": "Record identifier",
            "in": "path",
            "name": "outOfBandId",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/RecordId"
            }
          }
        ]
      }
    },
    "/didcomm/oob/create-invitation": {
      "post": {
        "operationId": "CreateInvitation",
        "responses": {
          "200": {
            "description": "Out of band record",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "invitationDid": {
                      "type": "string"
                    },
                    "outOfBandRecord": {
                      "$ref": "#/components/schemas/Record_string.unknown_"
                    },
                    "invitation": {
                      "$ref": "#/components/schemas/PlaintextMessage"
                    },
                    "invitationUrl": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "invitationDid",
                    "outOfBandRecord",
                    "invitation",
                    "invitationUrl"
                  ],
                  "type": "object"
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "invitationUrl": "string",
                      "invitation": {
                        "@type": "did:sov:BzCbsNYhMrjHiqZDTUASHg;spec/out-of-band/1.1/invitation",
                        "@id": "d6472943-e5d0-4d95-8b48-790ed5a41931",
                        "label": "Aries Test Agent",
                        "accept": [
                          "didcomm/aip1",
                          "didcomm/aip2;env=rfc19"
                        ],
                        "handshake_protocols": [
                          "https://didcomm.org/didexchange/1.0",
                          "https://didcomm.org/connections/1.0"
                        ],
                        "services": [
                          {
                            "id": "#inline-0",
                            "serviceEndpoint": "https://6b77-89-20-162-146.ngrok.io",
                            "type": "did-communication",
                            "recipientKeys": [
                              "did:key:z6MkmTBHTWrvLPN8pBmUj7Ye5ww9GiacXCYMNVvpScSpf1DM"
                            ],
                            "routingKeys": []
                          }
                        ]
                      },
                      "outOfBandRecord": {
                        "_tags": {
                          "invitationId": "1cbd22e4-1906-41e9-8807-83d84437f978",
                          "state": "await-response",
                          "role": "sender",
                          "recipientKeyFingerprints": [
                            "z6MktUCPZjfRJXD4GMcYuXiqX2qZ8vBw6UAYpDFiHEUfwuLj"
                          ]
                        },
                        "outOfBandInvitation": {
                          "@type": "did:sov:BzCbsNYhMrjHiqZDTUASHg;spec/out-of-band/1.1/invitation",
                          "@id": "d6472943-e5d0-4d95-8b48-790ed5a41931",
                          "label": "Aries Test Agent",
                          "accept": [
                            "didcomm/aip1",
                            "didcomm/aip2;env=rfc19"
                          ],
                          "handshake_protocols": [
                            "https://didcomm.org/didexchange/1.0",
                            "https://didcomm.org/connections/1.0"
                          ],
                          "services": [
                            {
                              "id": "#inline-0",
                              "serviceEndpoint": "https://6b77-89-20-162-146.ngrok.io",
                              "type": "did-communication",
                              "recipientKeys": [
                                "did:key:z6MkmTBHTWrvLPN8pBmUj7Ye5ww9GiacXCYMNVvpScSpf1DM"
                              ],
                              "routingKeys": []
                            }
                          ]
                        },
                        "metadata": {},
                        "id": "42a95528-0e30-4f86-a462-0efb02178b53",
                        "createdAt": "2022-01-01T00:00:00.000Z",
                        "reusable": false
                      }
                    }
                  }
                }
              }
            },
            "headers": {
              "x-correlation-id": {
                "description": "Correlation ID for tracing across systems",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "description": "Creates an outbound out-of-band record containing out-of-band invitation message defined in\nAries RFC 0434: Out-of-Band Protocol 1.1.",
        "tags": [
          "DIDComm - Out Of Band"
        ],
        "security": [
          {
            "jwt": [
              "tenant",
              "dedicated"
            ]
          }
        ],
        "parameters": [],
        "requestBody": {
          "description": "configuration of how out-of-band invitation should be created",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/CreateInvitationOptions"
                  },
                  {
                    "$ref": "#/components/schemas/RecipientKeyOption"
                  }
                ],
                "description": "configuration of how out-of-band invitation should be created"
              }
            }
          }
        }
      }
    },
    "/didcomm/oob/create-legacy-invitation": {
      "post": {
        "operationId": "CreateLegacyInvitation",
        "responses": {
          "200": {
            "description": "out-of-band record and invitation",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "outOfBandRecord": {
                      "$ref": "#/components/schemas/Record_string.unknown_"
                    },
                    "invitation": {
                      "$ref": "#/components/schemas/PlaintextMessage"
                    },
                    "invitationUrl": {
                      "type": "string"
                    },
                    "recipientKey": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "outOfBandRecord",
                    "invitation",
                    "invitationUrl"
                  ],
                  "type": "object"
                },
                "examples": {
                  "Example 1": {
                    "value": {
                      "invitation": {
                        "@type": "did:sov:BzCbsNYhMrjHiqZDTUASHg;spec/out-of-band/1.1/invitation",
                        "@id": "d6472943-e5d0-4d95-8b48-790ed5a41931",
                        "label": "Aries Test Agent",
                        "accept": [
                          "didcomm/aip1",
                          "didcomm/aip2;env=rfc19"
                        ],
                        "handshake_protocols": [
                          "https://didcomm.org/didexchange/1.0",
                          "https://didcomm.org/connections/1.0"
                        ],
                        "services": [
                          {
                            "id": "#inline-0",
                            "serviceEndpoint": "https://6b77-89-20-162-146.ngrok.io",
                            "type": "did-communication",
                            "recipientKeys": [
                              "did:key:z6MkmTBHTWrvLPN8pBmUj7Ye5ww9GiacXCYMNVvpScSpf1DM"
                            ],
                            "routingKeys": []
                          }
                        ]
                      },
                      "outOfBandRecord": {
                        "_tags": {
                          "invitationId": "1cbd22e4-1906-41e9-8807-83d84437f978",
                          "state": "await-response",
                          "role": "sender",
                          "recipientKeyFingerprints": [
                            "z6MktUCPZjfRJXD4GMcYuXiqX2qZ8vBw6UAYpDFiHEUfwuLj"
                          ]
                        },
                        "outOfBandInvitation": {
                          "@type": "did:sov:BzCbsNYhMrjHiqZDTUASHg;spec/out-of-band/1.1/invitation",
                          "@id": "d6472943-e5d0-4d95-8b48-790ed5a41931",
                          "label": "Aries Test Agent",
                          "accept": [
                            "didcomm/aip1",
                            "didcomm/aip2;env=rfc19"
                          ],
                          "handshake_protocols": [
                            "https://didcomm.org/didexchange/1.0",
                            "https://didcomm.org/connections/1.0"
                          ],
                          "services": [
                            {
                              "id": "#inline-0",
                              "serviceEndpoint": "https://6b77-89-20-162-146.ngrok.io",
                              "type": "did-communication",
                              "recipientKeys": [
                                "did:key:z6MkmTBHTWrvLPN8pBmUj7Ye5ww9GiacXCYMNVvpScSpf1DM"
                              ],
                              "routingKeys": []
                            }
                          ]
                        },
                        "metadata": {},
                        "id": "42a95528-0e30-4f86-a462-0efb02178b53",
                        "createdAt": "2022-01-01T00:00:00.000Z",
                        "reusable": false
                      }
                    }
                  }
                }
              }
            },
            "headers": {
              "x-correlation-id": {
                "description": "Correlation ID for tracing across systems",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "description": "Creates an outbound out-of-band record in the same way how `createInvitation` method does it,\nbut it also converts out-of-band invitation message to an \"legacy\" invitation message defined\nin RFC 0160: Connection Protocol and returns it together with out-of-band record.",
        "tags": [
          "DIDComm - Out Of Band"
        ],
        "security": [
          {
            "jwt": [
              "tenant",
              "dedicated"
            ]
          }
        ],
        "parameters": [],
        "requestBody": {
          "description": "configuration of how a invitation should be created",
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/Omit_CreateLegacyInvitationConfig.routing_"
                  },
                  {
                    "$ref": "#/components/schemas/RecipientKeyOption"
                  }
                ],
                "description": "configuration of how a invitation should be created"
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}